Create new file for blogFollower
